Feat: 마커 기능 구현 (#10)
* Feat: 마커 더미 데이터 생성 로직 구현 #8

* Feat: 마커 줍기 기능 구현 #8

* Feat: 마커 불러오기 기능 구현 #8

* Feat: 예외 처리 설정

* Style: 탭 컨벤션 수정

+++ src/main/java/com/example/travelhana/Projection/AccountInfoProjection.java
@@ -2,7 +2,7 @@ package com.example.travelhana.Projection;
 
 public interface AccountInfoProjection {
 
-	Long getId();
+	int getId();
 	String getAccountNum();
 	String getBank();
 	Long getBalance();
